chore: Rename follow.dto.ts
